requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.4
langdetect==1.0.9
requests==2.31.0